Added Spring dependency
