--- src/main/java/entity/CustomList.java
@@ -1,11 +1,0 @@
-package entity;
-
-import model.ticket.Ticket;
-
-public interface CustomList {
-    void put(Ticket ticket);
-
-    Ticket get(int index);
-
-    boolean delete(int index);
-}
