--- src/main/java/entity/Printable.java
@@ -1,8 +1,0 @@
-package entity;
-
-public interface Printable {
-
-    default void print() {
-        System.out.println("print content in console");
-    }
-}
