--- src/main/java/entity/Shareable.java
@@ -1,9 +1,0 @@
-package entity;
-
-public interface Shareable {
-    void share();
-
-    void share(String phone);
-
-    void share(String phone, String email);
-}
