--- src/main/java/model/StadiumSector.java
@@ -1,7 +1,0 @@
-package model;
-
-public enum StadiumSector {
-    A,
-    B,
-    C
-}
