--- src/main/java/model/TicketType.java
@@ -1,8 +1,0 @@
-package model;
-
-public enum TicketType {
-    DAY,
-    WEEK,
-    MONTH,
-    YEAR
-}
