--- src/main/java/model/ticket/BusTicket.java
@@ -1,9 +1,0 @@
-package model.ticket;
-
-public class BusTicket extends Ticket {
-
-    @Override
-    public void share(String phone) {
-        System.out.println("Send SMS to:" + phone);
-    }
-}
