+++ src/main/java/org/example/entity/CustomList.java
@@ -1,0 +1,11 @@
+package org.example.entity;
+
+import org.example.model.ticket.Ticket;
+
+public interface CustomList {
+    void put(Ticket ticket);
+
+    Ticket get(int index);
+
+    boolean delete(int index);
+}
