+++ src/main/java/org/example/entity/CustomSet.java
@@ -1,0 +1,14 @@
+package org.example.entity;
+
+import org.example.model.ticket.Ticket;
+
+public interface CustomSet {
+    boolean put(Ticket ticket);
+
+    boolean check(Ticket ticket);
+
+    void iterate();
+
+    void delete(Ticket ticket);
+
+}
