+++ src/main/java/org/example/entity/Printable.java
@@ -1,0 +1,8 @@
+package org.example.entity;
+
+public interface Printable {
+
+    default void print() {
+        System.out.println("print content in console");
+    }
+}
