+++ src/main/java/org/example/entity/Shareable.java
@@ -1,0 +1,9 @@
+package org.example.entity;
+
+public interface Shareable {
+    void share();
+
+    void share(String phone);
+
+    void share(String phone, String email);
+}
