+++ src/main/java/org/example/model/StadiumSector.java
@@ -1,0 +1,7 @@
+package org.example.model;
+
+public enum StadiumSector {
+    A,
+    B,
+    C
+}
