+++ src/main/java/org/example/model/TicketType.java
@@ -1,0 +1,8 @@
+package org.example.model;
+
+public enum TicketType {
+    DAY,
+    WEEK,
+    MONTH,
+    YEAR
+}
