+++ src/main/java/org/example/model/ticket/BusTicket.java
@@ -1,0 +1,9 @@
+package org.example.model.ticket;
+
+public class BusTicket extends Ticket {
+
+    @Override
+    public void share(String phone) {
+        System.out.println("Send SMS to:" + phone);
+    }
+}
