+++ src/main/java/org/example/model/ticket/CinemaTicket.java
@@ -1,0 +1,9 @@
+package org.example.model.ticket;
+
+public class CinemaTicket extends Ticket {
+
+    @Override
+    public void share(String phone, String email) {
+        System.out.println("Send SMS and email:" + phone + "," + email);
+    }
+}
